Demote previous-gen propertyGenerator() from class.
Move to previous iteration unit test.
