+++ .vscode/settings.json
@@ -5,7 +5,6 @@
     "editorError.foreground": "#ffef0f",
     "editorWarning.foreground": "#3777ff"
 },
-  "phpstan.showProgress": true,
 "phpstan.rootDir": "",
 
   "phpcs.showSources": true,
@@ -13,5 +12,6 @@
   "phpcs.executablePath": "./vendor/bin/phpcs",
   "phpcs.enable": true,
   "editor.tabSize": 4,
-  "editor.detectIndentation": false,
+"editor.detectIndentation": false,
+"phpstan.enableLanguageServer": false,
 }
