+++ src/Bread/MediaNoche.php
@@ -26,50 +26,6 @@ final class MediaNoche
     }
 
     /**
-     * Recursive property generator for Nette.
-     *
-     * Creates nette class `Property` objects from cebe `Schema` objects.
-     * Uses a flatten/merge recursive pattern to retrieve nested object properties.
-     *
-     * @return \Generator<string, Property, null, void>
-     */
-    public function propertyGenerator(Schema $schema): \Generator
-    {
-        foreach ($schema->properties as $name => $property) {
-            $this->logger->debug(sprintf('Parsing property: %s', $name));
-
-            if ($property->type == 'object') {
-                // Start a new internal, recursive generator.
-                $this->logger->debug(sprintf('Recursing object property: %s', $name));
-                foreach ($this->propertyGenerator($property) as $key => $nette_prop) {
-                    yield $key => $nette_prop;
-                }
-                // Do not yield Schema items, only Property items.
-                return;
-            }
-
-            /* @see https://swagger.io/specification/#data-types */
-            $type = match ($property->type) {
-                'string' => 'string',
-                'integer' => 'int',
-                'boolean' => 'bool',
-                'float', 'double' => 'float',
-                // 'object' => Schema::class,
-                'date', 'dateTime' => \DateTimeInterface::class,
-                default => throw new \UnhandledMatchError(),
-            };
-
-            yield $name =>
-            (new Property($name))
-                ->setType($type)
-                ->setReadOnly(true)
-                ->setComment($property->description)
-                ->setNullable(true)
-                ->setValue($property->default);
-        }
-    }
-
-    /**
      * Converts a property from a cebe to a nette object.
      *
      * @param Schema $property
